:recycle: Move models into their own files
