+++ src/openpersonen/api/testing_models/__init__.py
@@ -1,1 +1,12 @@
-from .models import *  # noqa
+from .gezagsverhouding import *  # noqa
+from .inschrijving import *  # noqa
+from .kiesrecht import *  # noqa
+from .kind import *  # noqa
+from .nationaliteit import *  # noqa
+from .ouder import *  # noqa
+from .overlijden import *  # noqa
+from .partnerschap import *  # noqa
+from .persoon import *  # noqa
+from .reisdocument import *  # noqa
+from .verblijfplaats import *  # noqa
+from .verblijfstitel import *  # noqa
